feat: rewrite code in TypeScript and use new tailwind v3 plugin API
